update styled-types package all theme
