+++ packages/localize-components-styled-types/src/index.ts
@@ -1,2 +1,2 @@
+export * from './styles';
 export * from './theme';
-export * from './types';
