+++ packages/localize-components-styled-types/src/styles/LocalizeAlign.ts
@@ -1,0 +1,3 @@
+export type Align = 'center' | 'left' | 'right';
+
+export default Align;
