+++ packages/localize-components-styled-types/src/styles/LocalizePosition.ts
@@ -1,0 +1,12 @@
+export type Position =
+  | 'bottom-center'
+  | 'bottom-left'
+  | 'bottom-right'
+  | 'center'
+  | 'left'
+  | 'right'
+  | 'top-center'
+  | 'top-left'
+  | 'top-right';
+
+export default Position;
