+++ packages/localize-components-styled-types/src/styles/LocalizeSize.ts
@@ -1,0 +1,3 @@
+export type Size = 'large' | 'medium' | 'small';
+
+export default Size;
