+++ packages/localize-components-styled-types/src/styles/index.ts
@@ -1,0 +1,3 @@
+export * from './LocalizeAlign';
+export * from './LocalizePosition';
+export * from './LocalizeSize';
