--- packages/localize-components-styled-types/src/types/index.ts
@@ -1,1 +1,0 @@
-export * from './styles';
