--- packages/localize-components-styled-types/src/types/styles/Align.ts
@@ -1,7 +1,0 @@
-export enum Align {
-  CENTER = 'center',
-  LEFT = 'left',
-  RIGHT = 'right',
-}
-
-export type AlignType = Align | 'center' | 'left' | 'right';
