--- packages/localize-components-styled-types/src/types/styles/Position.ts
@@ -1,25 +1,0 @@
-export enum Position {
-  BC = 'bottom-center',
-  BL = 'bottom-left',
-  BR = 'bottom-right',
-  CENTER = 'center',
-  LEFT = 'left',
-  RIGHT = 'right',
-  TC = 'top-center',
-  TL = 'top-left',
-  TR = 'top-right',
-}
-
-export type PositionType =
-  | Position
-  | 'bottom-center'
-  | 'bottom-left'
-  | 'bottom-right'
-  | 'center'
-  | 'left'
-  | 'right'
-  | 'top-center'
-  | 'top-left'
-  | 'top-right';
-
-export default Position;
