--- packages/localize-components-styled-types/src/types/styles/Size.ts
@@ -1,9 +1,0 @@
-export enum Size {
-  LARGE = 'large',
-  MEDIUM = 'medium',
-  SMALL = 'small',
-}
-
-export type SizeType = Size | 'large' | 'medium' | 'small';
-
-export default Size;
