--- packages/localize-components-styled-types/src/types/styles/index.ts
@@ -1,6 +1,0 @@
-export * from './Align';
-export * from './FontSizes';
-export * from './Position';
-export * from './Size';
-export * from './Theme';
-export * from './GridTypes';
